Bodge for cygwin: link swrast.so with libGL.dll.a to resolve all symbols
_glapi_Dispatch/_glapi_Context are supposed to be defined by
the implementor (i.e. XWin).  The problem is that if there is no
single implementor, then what to resolve against?

We link against libGL.dll.a to provide those symbols (and take other
steps in XWin to avoid duplicating symbols there)

(For a xlib-driver libGL, this introduces a circular dependency, as
libGL is linked to swrast; not an issue for dri-driver libGL, as that
dynamically loads swrast...)

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>