Some fixes for GLX_INDIRECT_RENDERING only build
This fixes some of the build issues with GLX_INDIRECT_RENDERING but not GLX_DIRECT_RENDERING due to recent changes.

There still remain some issues with the use of psc->dpy in indirect_glx.c.
I'm not sure if there's more cleanup coming and that should change to psc->display->dpy?

Signed-off-by: Jon TURNEY <jon.turney@dronecode.org.uk>